feat: 移動している矩形同士の判定関数作成
ファイル分けに失敗しているので後で直すべき

+++ J6/GameManager.cpp
@@ -5,7 +5,7 @@ namespace MyGame{
 	class GameManager {
 	private:
 		Player* player;
-		int16 chipSizeInt;
+		int chipSizeInt;
 		Size chipSize;
 		Array<GameObject*> gameObjects;
 		Camera2D camera{ Vec2{ 0, 0 }, 1.0 };
